+++ tmanga/src/main/java/com/website/springmvc/controller/PublishCompanyAdminController.java
@@ -19,6 +19,7 @@ import org.springframework.web.servlet.ModelAndView;
 
 import com.website.springmvc.Services.PublishCompanyService;
 import com.website.springmvc.entities.PublishCompany;
+import com.website.springmvc.libs.GetModel;
 import com.website.springmvc.libs.RemoveAccent;;
 @Controller
 @RequestMapping(value = "/controller")
@@ -27,53 +28,65 @@ public class PublishCompanyAdminController {
 	@Autowired
 	private PublishCompanyService pubcomService;
 	
+	@Autowired
+	GetModel getModel;
+	
 	@RequestMapping(value = "/pubcomAdmin", method = RequestMethod.GET)
 	public ModelAndView getAuthorAdmin(@RequestParam(name = "mes", defaultValue = "") String mes,
 									@RequestParam(name = "alert", defaultValue = "") String alert,
-									@RequestParam(name = "p", defaultValue = "1") int page) {
+									@RequestParam(name = "p", defaultValue = "1") int page,
+									@RequestParam(name = "q", defaultValue = "") String key, HttpSession session) {
 		ModelAndView model = new ModelAndView();
-		model.setViewName("admin/layoutAdmin");
-		model.addObject("views","PublishingAdmin");
-		model.addObject("title","Danh sách nhà xuất bản");
+		boolean f = getModel.checkAdmin(session);
 		
-		List<PublishCompany> pubComs = pubcomService.getListPublishCompany(0,0);
-
-		int totalPage = 0;
-		int totalCate = 0;		
-
-		totalCate = pubComs.size();
-		totalPage = totalCate / 10;
-
-		if(totalCate % 10 != 0){
-			totalPage++;
-		}		
-
-		int start = 1, end = 7;
-
-		if(totalPage > 7){		
-			if(page - 3 > 0){	
-				if(page + 3 >= totalPage){	
-					start =  totalPage - 6;
-					end = totalPage;
+		if(f) {
+			getModel.getLayoutAdmin(model);
+			model.addObject("views","PublishingAdmin");
+			model.addObject("title","Danh sách nhà xuất bản");
+			
+			List<PublishCompany> pubComs = pubcomService.getListPublishCompany(key, 0, 0);
+	
+			int totalPage = 0;
+			int totalCate = 0;		
+	
+			totalCate = pubComs.size();
+			totalPage = totalCate / 10;
+	
+			if(totalCate % 10 != 0){
+				totalPage++;
+			}		
+	
+			int start = 1, end = 7;
+	
+			if(totalPage > 7){		
+				if(page - 3 > 0){	
+					if(page + 3 >= totalPage){	
+						start =  totalPage - 6;
+						end = totalPage;
+					}	
+					else{	
+						start = page - 3;
+						end = page + 3;
+					}	
 				}	
-				else{	
-					start = page - 3;
-					end = page + 3;
-				}	
-			}	
+			}
+			else {
+				end = totalPage;
+			}
+	
+			model.addObject("start", start);
+			model.addObject("end", end);
+	
+			pubComs = pubcomService.getListPublishCompany(key, 10*(page-1), 10);
+	
+			model.addObject("PubComs", pubComs);
+			model.addObject("totalpage", totalPage);
+			model.addObject("pageselected", page);
+			model.addObject("key", key);
 		}
 		else {
-			end = totalPage;
+			getModel.getHome(model, session);
 		}
-
-		model.addObject("start", start);
-		model.addObject("end", end);
-
-		pubComs = pubcomService.getListPublishCompany(10*(page-1), 10);
-
-		model.addObject("PubComs", pubComs);
-		model.addObject("totalpage", totalPage);
-		model.addObject("pageselected", page);
 		return model;
 	}
 	
@@ -83,29 +96,35 @@ public class PublishCompanyAdminController {
 							@RequestParam(name = "mode", defaultValue = "add") String mode,
 							@RequestParam(name = "id", defaultValue = "0") Long idPC){
 		String str = "redirect:pubcomAdmin";
-
-		if(mode.equalsIgnoreCase("add")) {
-			PublishCompany publishCompany = new PublishCompany();
-			publishCompany.setName(name);
-			publishCompany.setUnsignedName(RemoveAccent.changeTitle(name));
-			pubcomService.add(publishCompany);
-			model.addAttribute("mes", "Thêm thành công");
-			model.addAttribute("alert", "success");
-		}
-		else if(mode.equalsIgnoreCase("edit")) {
-			if(pubcomService.get(idPC) == null) {
-				model.addAttribute("mes", "ID không tồn tại");
-				model.addAttribute("alert", "danger");
-			}
-			else{
-				PublishCompany publishCompany = pubcomService.get(idPC);
+		boolean f = getModel.checkAdmin(session);
+		
+		if(f) {			
+			if(mode.equalsIgnoreCase("add")) {
+				PublishCompany publishCompany = new PublishCompany();
 				publishCompany.setName(name);
 				publishCompany.setUnsignedName(RemoveAccent.changeTitle(name));
-				pubcomService.update(publishCompany);
-				model.addAttribute("mes", "Sửa thành công");
+				pubcomService.add(publishCompany);
+				model.addAttribute("mes", "Thêm thành công");
 				model.addAttribute("alert", "success");
 			}
-		}		
+			else if(mode.equalsIgnoreCase("edit")) {
+				if(pubcomService.get(idPC) == null) {
+					model.addAttribute("mes", "ID không tồn tại");
+					model.addAttribute("alert", "danger");
+				}
+				else{
+					PublishCompany publishCompany = pubcomService.get(idPC);
+					publishCompany.setName(name);
+					publishCompany.setUnsignedName(RemoveAccent.changeTitle(name));
+					pubcomService.update(publishCompany);
+					model.addAttribute("mes", "Sửa thành công");
+					model.addAttribute("alert", "success");
+				}
+			}
+		}
+		else {
+			str = "redirect:index";
+		}
 		return str;
 	}
 
@@ -113,30 +132,37 @@ public class PublishCompanyAdminController {
 	public String showHide(HttpSession session, Model model,								
 							@RequestParam(name = "id", defaultValue = "0") Long idPC) {
 		String str = "redirect:pubcomAdmin";
-		if(pubcomService.get(idPC) != null) {
-			PublishCompany publishCompany = pubcomService.get(idPC);
-			if(publishCompany.getStatus() == 0){
-				publishCompany.setStatus((byte) 1);
+		boolean f = getModel.checkAdmin(session);
+		
+		if(f) {	
+			if(pubcomService.get(idPC) != null) {
+				PublishCompany publishCompany = pubcomService.get(idPC);
+				if(publishCompany.getStatus() == 0){
+					publishCompany.setStatus((byte) 1);
+				}
+				else if(publishCompany.getStatus() == 1){
+					publishCompany.setStatus((byte) 0);
+				}
+				pubcomService.update(publishCompany);
+				model.addAttribute("mes", "Cập nhật thành công");
+				model.addAttribute("alert", "success");
 			}
-			else if(publishCompany.getStatus() == 1){
-				publishCompany.setStatus((byte) 0);
+			else{
+				model.addAttribute("mes", "ID không tồn tại");
+				model.addAttribute("alert", "danger");
 			}
-			pubcomService.update(publishCompany);
-			model.addAttribute("mes", "Cập nhật thành công");
-			model.addAttribute("alert", "success");
 		}
-		else{
-			model.addAttribute("mes", "ID không tồn tại");
-			model.addAttribute("alert", "danger");
+		else {
+			str = "redirect:index";
 		}
 		return str;
 	}
 
 	@RequestMapping(value = "/removePublishing", method = RequestMethod.POST)
-	public void removeAddress(@RequestParam(name = "id") Long idPC, HttpServletResponse response) {
+	public void removeAddress(@RequestParam(name = "id") Long idPC, HttpServletResponse response, HttpSession session) {
 		String str = "success";
-
-		if(pubcomService.get(idPC) != null) {
+		boolean f = getModel.checkAdmin(session);
+		if(f && pubcomService.get(idPC) != null) {
 			pubcomService.delete(idPC);
 		}
 		else {
